Refactor Systems exposed table toolbar
